backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
anthropic==0.25.1
pdfplumber==0.10.3
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0 